# new_project/contracts/openapi/investment-api.yaml
openapi: 3.1.0
info:
  title: Personal Investment API
  version: 0.2.0
  description: 本地开发切片。所有金额以原币种两位小数最小单位的十进制字符串传输，绝不使用浮点数。
servers:
  - url: http://127.0.0.1:<API_PORT>
paths:
...
  /api/v1/ledger/accounts:
    get:
      summary: 查询当前管理员的现金账户
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 现金账户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashAccountListResponse'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: 创建现金账户并初始化同币种后端系统科目
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCashAccountRequest'
      responses:
        '201':
          description: 已创建；同键同请求返回首个响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashAccount'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /api/v1/market/snapshot-submissions:
    post:
      summary: 提交经核验的市场快照；仅入队，不在请求线程调用外部数据源
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MarketSnapshotSubmissionRequest'}
      responses:
        '202':
          description: 已持久化并等待后台处理
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MarketSnapshotSubmissionAccepted'}
        '422': {$ref: '#/components/responses/Problem'}
...
  /api/v1/ledger/accounts/{accountId}/disable:
    post:
      summary: 停用现金账户
      description: 仅余额为零、没有现货/期货持仓、保证金余额及活动导入引用时可停用；不删除历史事实。
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
            pattern: '^"?[0-9]+"?$'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      responses:
        '200':
          description: 已停用并递增版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashAccount'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    Problem:
      description: RFC 7807 风格错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateCashAccountRequest:
      type: object
      required: [displayName, currency]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        currency:
          type: string
          enum: [CNY, USD]
    CashAccount:
      type: object
      required: [accountId, displayName, accountKind, currency, status, version]
      properties:
        accountId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        displayName:
          type: string
        accountKind:
          type: string
          enum: [ASSET_CASH]
        currency:
          type: string
          enum: [CNY, USD]
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        version:
          type: string
          pattern: '^[0-9]+$'
    CashAccountListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CashAccount'
...
    MarketSnapshotSubmissionRequest:
      type: object
      required: [tradingDate, sourceName, sourceReference]
      properties:
        tradingDate: {type: string, format: date}
        sourceName: {type: string, maxLength: 64}
        sourceReference: {type: string, maxLength: 512}
        quotes:
          type: array
          items: {$ref: '#/components/schemas/MarketQuoteSubmission'}
        metrics:
          type: array
          items: {$ref: '#/components/schemas/MarketMetricSubmission'}
        basis:
          type: array
          items: {$ref: '#/components/schemas/MarketBasisSubmission'}
    MarketQuoteSubmission:
      type: object
      required: [instrumentId, quoteTime, sourceObservationKey, priceCent, currency]
      properties:
        instrumentId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        quoteTime: {type: string, format: date-time}
        sourceObservationKey: {type: string, maxLength: 256}
        priceCent: {type: string, pattern: '^[1-9][0-9]*$'}
        prevCloseCent: {type: [string, 'null'], pattern: '^[1-9][0-9]*$'}
        currency: {type: string, enum: [CNY, USD]}
    MarketMetricSubmission:
      type: object
      required: [instrumentId, metricName, metricValueDecimal, sourceObservationKey]
      properties:
        instrumentId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        metricName: {type: string, maxLength: 64}
        metricValueDecimal: {type: string}
        sourceObservationKey: {type: string, maxLength: 256}
    MarketBasisSubmission:
      type: object
      required: [underlyingInstrumentId, futureInstrumentId, spotPricePoints, futurePricePoints, sourceObservationKey]
      properties:
        underlyingInstrumentId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        futureInstrumentId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        spotPricePoints: {type: string}
        futurePricePoints: {type: string}
        annualizedBasisDecimal: {type: [string, 'null']}
        maturityDate: {type: [string, 'null'], format: date}
        daysLeft: {type: [integer, 'null'], minimum: 0}
        sourceObservationKey: {type: string, maxLength: 256}
    MarketSnapshotSubmissionAccepted:
      type: object
      required: [marketSnapshotSubmissionId, marketSyncRunId, status]
      properties:
        marketSnapshotSubmissionId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        marketSyncRunId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        status: {type: string, enum: [QUEUED]}
...
    Problem:
      type: object
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          type: array
          items:
            type: string